% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extractPatients}
\alias{extractPatients}
\title{ExtractPatients from a connection}
\usage{
extractPatients(connection)
}
\arguments{
\item{connection}{connection to the Common Data Model (CDM)}
}
\value{
data.frame with count of condition concepts ID over year and month.
}
\description{
ExtractPatients from a connection
}
