% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShinyApp.R
\name{launchShinyApp}
\alias{launchShinyApp}
\title{Launch app}
\usage{
launchShinyApp(data)
}
\arguments{
\item{data}{data taked from the environment}
}
\description{
This function launch a shiny-based web app to display a graph of condition count
}
