% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot of condition counts}
\usage{
plotTrend(data, byMonth = FALSE)
}
\arguments{
\item{data}{data.frame with count of condition concepts ID over year and month.}

\item{byMonth}{Boolean display the counts by month?}
}
\value{
ggplotly with count of conditions by month or by year
}
\description{
Plot of condition counts
}
